Create Calculate the nth term
In C programming, calculating the $n^{th}$ term usually refers to solving a sequence using recursion. A common version of this challenge (found on HackerRank) asks you to find the $n^{th}$ term of a sequence where each term is the sum of the three previous terms.This is very similar to the Fibonacci sequence, but instead of $fib(n) = fib(n-1) + fib(n-2)$, it uses three terms.